# backend/docs/swagger.yaml
basePath: /
definitions:
  domain.Role:
    enum:
    - admin
    - coach
    - player
    - parent
    - analyst
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleCoach
    - RolePlayer
    - RoleParent
    - RoleAnalyst
  dto.AuthResponse:
    properties:
      accessToken:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.ClubDashboardResponse:
    properties:
      club:
        $ref: '#/definitions/dto.ClubResponse'
      coachCount:
        type: integer
      playerCount:
        type: integer
      recentSessions:
        items:
          $ref: '#/definitions/dto.SessionResponse'
        type: array
      teamCount:
        type: integer
      teamStats:
        items:
          $ref: '#/definitions/dto.TeamStatItem'
        type: array
    type: object
  dto.ClubResponse:
    properties:
      city:
        type: string
      country:
        type: string
      createdAt:
        type: string
      foundedAt:
        type: string
      id:
        type: string
      logoUrl:
        type: string
      name:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.MatchResponse:
    properties:
      createdAt:
        type: string
      goalsAgainst:
        type: integer
      goalsFor:
        type: integer
      id:
        type: string
      isHome:
        type: boolean
      kickoffAt:
        type: string
      location:
        type: string
      notes:
        type: string
      opponent:
        type: string
      status:
        type: string
      teamId:
        type: string
    type: object
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      clubID:
        type: string
      email:
        maxLength: 255
        type: string
      firstName:
        maxLength: 100
        minLength: 1
        type: string
      lastName:
        maxLength: 100
        minLength: 1
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      role:
        $ref: '#/definitions/domain.Role'
    required:
    - email
    - firstName
    - lastName
    - password
    - role
    type: object
  dto.SaveBlockExerciseItem:
    properties:
      durationMin:
        type: integer
      exerciseId:
        type: string
      intensityOverride:
        type: string
      orderIndex:
        type: integer
      reps:
        type: integer
      sets:
        type: integer
    required:
    - exerciseId
    type: object
  dto.SaveBlockItem:
    properties:
      durationMin:
        type: integer
      exercises:
        items:
          $ref: '#/definitions/dto.SaveBlockExerciseItem'
        type: array
      kind:
        enum:
        - warmup
        - main
        - game
        - cooldown
        type: string
      notes:
        type: string
      orderIndex:
        type: integer
    required:
    - kind
    type: object
  dto.SaveBlocksRequest:
    properties:
      blocks:
        items:
          $ref: '#/definitions/dto.SaveBlockItem'
        type: array
    required:
    - blocks
    type: object
  dto.SessionResponse:
    properties:
      coachId:
        type: string
      createdAt:
        type: string
      durationMin:
        type: integer
      focus:
        items:
          type: string
        type: array
      id:
        type: string
      intensity:
        type: string
      location:
        type: string
      notes:
        type: string
      scheduledAt:
        type: string
      status:
        type: string
      teamId:
        type: string
    type: object
  dto.TeamStatItem:
    properties:
      avgAttendance:
        type: number
      avgDevIndex:
        type: number
      playerCount:
        type: integer
      teamId:
        type: string
      teamName:
        type: string
    type: object
  dto.UserResponse:
    properties:
      avatarURL:
        type: string
      clubId:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      lastLoginAt:
        type: string
      lastName:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for Football Manager system
  title: CoachOS API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return tokens
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Returns profile of authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - auth
  /api/v1/clubs/{id}/dashboard:
    get:
      description: Returns club statistics and overview
      parameters:
      - description: Club ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClubDashboardResponse'
      security:
      - BearerAuth: []
      summary: Get club dashboard
      tags:
      - clubs
  /api/v1/matches:
    get:
      description: Returns paginated list of matches
      parameters:
      - description: Match status filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MatchResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List matches
      tags:
      - matches
  /api/v1/sessions/{id}/blocks:
    put:
      consumes:
      - application/json
      description: Atomically replaces all blocks and exercises for a session.
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Blocks payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SaveBlocksRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Save session blocks
      tags:
      - sessions
  /health:
    get:
      description: Returns API health status and version
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - system
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"